The final changes have been made, comments have been added to the code, and the responsiveness of the layout has been implemented. Implemented Guard, and much more.

+++ src/app/shared/base.service.ts
@@ -1,6 +1,7 @@
 import { MainService } from './main.service';
 import { Injectable } from '@angular/core';
-import { AngularFirestore, AngularFirestoreDocument } from '@angular/fire/firestore';
+import { AngularFirestore } from '@angular/fire/firestore';
+import { AuthService } from './auth.service';
 
 @Injectable({
   providedIn: 'root'
@@ -8,26 +9,23 @@ import { AngularFirestore, AngularFirestoreDocument } from '@angular/fire/firest
 export class BaseService {
 
   constructor(public afs: AngularFirestore,
-              private mainService: MainService) { }
+              private mainService: MainService,
+              public authService: AuthService) { }
 
-  getBookmarks(){
-    return this.afs.collection('bookmarks').snapshotChanges();
+  // Функция, возвращающая Observable на коллекцию Bookmar[], содержащую обьекты, сохранные пользователем авторизированным в данный момент.
+  getBookmarks(uid){
+    return this.afs.collection('bookmarks', ref => ref.where('uid', '==', uid)).snapshotChanges();
   }
 
+  // Добавляет в коллеккция bookmarks новый обьект.
   setBookmark(data) {
-    this.afs.collection("bookmarks").add(data)
+     this.afs.collection("bookmarks").add(data);
   }
 
-
-  deleteBookmark(id) {
-    let photoId: string;
-    this.afs.collection('bookmarks').snapshotChanges().subscribe(data => {
-      data.map(e => {
-        photoId = e.payload.doc.id
-
-      })
-    });
-    const deleteUserSubs = this.afs.collection('bookmarks').doc(id).delete();
+  // Принимает uid пользователя и id документа firestoreDB, передавая их метод, удаляющий запись из firestoreDB, где автор сохраненной
+  // записи в данный момент авторизированный пользователь.
+  deleteBookmark(uid, id) {
+    const deleteUserSubs = this.afs.collection('bookmarks', ref => ref.where('uid', '==', uid)).doc(id).delete();
     this.mainService.openSnackBar('photo successfully deleted!', "right", "bottom", "success-dialog-red")
   }
